Address comment to do SHJ codegen in doConsume as well
